fix(test): clear on teardown
